BUG#231795, NEWMS00231795, when play one long time mp4 file, if do seek, the video will stop and audio is ok.

+++ media/libstagefright/MPEG4Extractor.cpp
@@ -335,14 +335,14 @@ sp<MetaData> MPEG4Extractor::getTrackMetaData(
         CHECK(track->meta->findCString(kKeyMIMEType, &mime));
         if (!strncasecmp("video/", mime, 6)) {
             uint32_t sampleIndex;
-            uint32_t sampleTime;
+            uint64_t sampleTime;
             if (track->sampleTable->findThumbnailSample(&sampleIndex) == OK
                     && track->sampleTable->getMetaDataForSample(
                         sampleIndex, NULL /* offset */, NULL /* size */,
                         &sampleTime) == OK) {
                 track->meta->setInt64(
                         kKeyThumbnailTime,
-                        ((int64_t)sampleTime * 1000000) / track->timescale);
+                        (sampleTime * 1000000) / track->timescale);
             }
         }
     }
@@ -1596,7 +1596,7 @@ status_t MPEG4Source::read(
             return err;
         }
 
-        uint32_t sampleTime;
+        uint64_t sampleTime;
         CHECK_EQ((status_t)OK, mSampleTable->getMetaDataForSample(
                     sampleIndex, NULL, NULL, &sampleTime));
 
@@ -1627,7 +1627,7 @@ status_t MPEG4Source::read(
 
     off_t offset;
     size_t size;
-    uint32_t dts;
+    uint64_t dts;
     bool isSyncSample;
     bool newBuffer = false;
     if (mBuffer == NULL) {
